requirements.txt
setuptools~=62.1.0
tqdm~=4.64.0
numpy~=1.22.3
psutil~=5.9.0
redis~=4.2.2
scrapy~=2.6.1
requests~=2.27.1
urllib3~=1.26.9
pygame~=2.1.2
pillow~=9.1.0